+++ custom/plugins/SwagCustomizedProducts/src/Resources/app/administration/test/e2e/cypress/integration/customized-products.spec.js
@@ -7,6 +7,12 @@ const selector = {
         description: '.sw-text-editor__content',
         optionName: '#sw-field--newOption-displayName',
         optionType: '#sw-field--newOption-type',
+        numberfield: {
+            min: '#sw-field--option-typeProperties-minValue',
+            max: '#sw-field--option-typeProperties-maxValue',
+            step: '#sw-field--option-typeProperties-interval',
+            defaultValue: '#sw-field--option-typeProperties-defaultValue'
+        }
     },
     button: {
         addTemplate: '[href="#/swag/customized/products/create"]',
@@ -30,20 +36,8 @@ describe('Administration: Custom products', () => {
             });
     });
 
-    function assertListIsVisible() {
-        cy.get('.swag-customized-products-list').should('exist');
-    }
-
-    function createTemplate() {
-        cy.get(selector.button.addTemplate).click();
-        cy.get(selector.input.internalName).type('lorem-ipsum');
-        cy.get(selector.input.displayName).type('Lorem ipsum');
-        cy.get(selector.input.description).type('Lorem ipsum dolor sit amet...');
-        cy.get(selector.button.saveTemplate).click();
-    }
-
     it('@package @general: can navigate to custom products module', () => {
-        assertListIsVisible();
+        cy.get('.swag-customized-products-list').should('exist');
     });
 
     it('@package @general: can add a new template', () => {
@@ -68,6 +62,7 @@ describe('Administration: Custom products', () => {
 
         cy.get(selector.emptyState).should('not.exist');
 
+        // Add checkbox
         cy.contains('Add option').click();
         cy.get(selector.input.optionName).type('Check this');
         cy.get(selector.input.optionType).select('Checkbox');
@@ -75,7 +70,27 @@ describe('Administration: Custom products', () => {
         cy.get(selector.button.addOption).click();
         cy.get(selector.button.applyOption).click();
 
-        cy.get(selector.emptyState).should('exist');
+
+        // Add an additional numberfield
+        cy.contains('Add option').click();
+        cy.get(selector.input.optionName).type('Number this');
+        cy.get(selector.input.optionType).select('Number field');
+        cy.get(selector.button.addOption).click();
+
+        // Fill in typeProperties
+        cy.get(selector.input.numberfield.min).clear().type('2');
+        cy.get(selector.input.numberfield.max).clear().type('22');
+        cy.get(selector.input.numberfield.step).clear().type('2');
+        cy.get(selector.input.numberfield.defaultValue).clear().type('12');
+        cy.get(selector.button.applyOption).click();
+
+        // Reopen to check values
+        cy.contains('.sw-data-grid__cell-content', 'Number this').click();
+        cy.get(selector.input.numberfield.min).should('have.value', '2');
+        cy.get(selector.input.numberfield.max).should('have.value', '22');
+        cy.get(selector.input.numberfield.step).should('have.value', '2');
+        cy.get(selector.input.numberfield.defaultValue).should('have.value', '12');
+        cy.get(selector.button.cancelOption).click();
     });
 
     it('@package @general: can add a new template and try to add an option with the empty state remaining present', () => {