+++ custom/plugins/SwagCustomizedProducts/src/Resources/app/storefront/src/swag-customized-products-form-validator/swag-customized-products-form-validator.plugin.js
@@ -5,7 +5,10 @@ import DomAccess from 'src/helper/dom-access.helper';
 export default class SwagCustomizedProductsFormValidator extends Plugin {
     static options = {
         inputFieldsSelector: 'input, textarea',
-        buyButtonSelector: '#productDetailPageBuyProductForm .btn-buy'
+        selectors: {
+            buyButton: '#productDetailPageBuyProductForm .btn-buy',
+            fileUploadContainer: '.customized-products-upload'
+        }
     };
 
     init() {
@@ -39,7 +42,7 @@ export default class SwagCustomizedProductsFormValidator extends Plugin {
      * @private
      */
     _onFormChange() {
-        const buyButton = DomAccess.querySelector(this.parentEl, this.options.buyButtonSelector);
+        const buyButton = DomAccess.querySelector(this.parentEl, this.options.selectors.buyButton);
         const inputFields = DomAccess.querySelectorAll(this.parentEl, this.options.inputFieldsSelector, false);
         const selectionValidation = {};
         const requiredSimpleFieldsFilled = Array.from(inputFields).reduce((acc, field) => {
@@ -51,6 +54,28 @@ export default class SwagCustomizedProductsFormValidator extends Plugin {
                     field.removeAttribute('required');
                 }
             }
+
+            const fileUploadContainer = field.closest(this.options.selectors.fileUploadContainer);
+            // If we are dealing with a file upload
+            if (fileUploadContainer) {
+                // FileUpload not required
+                if (!DomAccess.getDataAttribute(field, 'required', false)) {
+                    return acc;
+                }
+
+                const fileUploadPlugin = window.PluginManager.getPluginInstanceFromElement(
+                    fileUploadContainer,
+                    'SwagCustomizedProductsFileUpload'
+                );
+                // The registry property holds the actual files and the input field value gets reset
+                if (fileUploadPlugin !== undefined && fileUploadPlugin.registry.size > 0) {
+                    return acc;
+                }
+
+                acc = false;
+                return acc;
+            }
+
 
             if (!acc || !field.required) {
                 return acc;